Allow maxSkew: 2 to encourage pods to schedule on 2 infra nodes
topologySpreadConstraints encourages pods to schedule one of the
monitoring pods on an infra node and one of them "not" on an
infra node. With maxSkew: 1, only one pod is allowed to be on an infra
node, otherwise if two pods were on an infra node and zero otherwise,
the "skew" would be 2.

With maxSkew: 2, we allow for a skew of 2, which results in both pods
scheduling on infra nodes.

This is a hack because the cluster monitoring operator does
not yet support nodeAffinity, which would solve this problem in a
clearer fashion.

Signed-off-by: Michael Shen <mshen@redhat.com>